backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dateutil==2.8.2
pytest==8.0.2
black==24.2.0
flake8==7.0.0 